+++ bundles/core/src/main/java/com/adobe/cq/wcm/core/components/internal/models/v2/TeaserImpl.java
@@ -17,6 +17,7 @@ package com.adobe.cq.wcm.core.components.internal.models.v2;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
@@ -33,6 +34,7 @@ import org.apache.sling.models.annotations.Model;
 import org.apache.sling.models.annotations.injectorspecific.ScriptVariable;
 import org.apache.sling.models.annotations.injectorspecific.Self;
 import org.apache.sling.models.annotations.injectorspecific.SlingObject;
+import org.apache.sling.settings.SlingSettingsService;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,30 +44,31 @@ import com.adobe.cq.export.json.ExporterConstants;
 import com.adobe.cq.wcm.core.components.internal.Utils;
 import com.adobe.cq.wcm.core.components.models.ListItem;
 import com.adobe.cq.wcm.core.components.models.Teaser;
+import com.day.cq.commons.Externalizer;
 import com.day.cq.wcm.api.Page;
 import com.day.cq.wcm.api.PageManager;
 import com.day.cq.wcm.api.components.Component;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
-@Model(adaptables = SlingHttpServletRequest.class, adapters = {Teaser.class, ComponentExporter.class}, resourceType = TeaserImpl.RESOURCE_TYPE)
-@Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME , extensions = ExporterConstants.SLING_MODEL_EXTENSION)
+@Model(adaptables = SlingHttpServletRequest.class, adapters = { Teaser.class,
+        ComponentExporter.class }, resourceType = TeaserImpl.RESOURCE_TYPE)
+@Exporter(name = ExporterConstants.SLING_MODEL_EXPORTER_NAME, extensions = ExporterConstants.SLING_MODEL_EXTENSION)
 public class TeaserImpl extends com.adobe.cq.wcm.core.components.internal.models.v1.TeaserImpl implements Teaser {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(TeaserImpl.class);
 
     public final static String RESOURCE_TYPE = "core/wcm/components/teaser/v2/teaser";
     public final static String OBJECT_ID = "s_objectID=";
-    public final static String SPACE = StringUtils.SPACE;
     public final static String APOSTROPHE = "'";
-    public final static String SEMICOLON = ";";   
-    
+    public final static String SEMICOLON = ";";
+
     private String compHashCode = StringUtils.EMPTY;
     private boolean trackingEnabled = false;
-    private StringBuilder analyticData ;
+    private StringBuilder linkTrackingCode;
     private int counter = 0;
-    String id = StringUtils.EMPTY;
-    
+    private String hashCodeValue = StringUtils.EMPTY;
     private List<ListItem> actions = new ArrayList<>();
+    private Set<String> runmode;
 
     @ScriptVariable
     private Component component;
@@ -75,33 +78,39 @@ public class TeaserImpl extends com.adobe.cq.wcm.core.components.internal.models
 
     @Inject
     private Resource resource;
-    
+
     @ScriptVariable
     private PageManager pageManager;
-   
+
     @Self
     private SlingHttpServletRequest request;
-    
+
     @SlingObject
-    private ResourceResolver resourceResolver;	
-    
+    private ResourceResolver resourceResolver;
+
     @ScriptVariable
-    protected com.day.cq.wcm.api.Page currentPage;   
+    protected Page currentPage;
+
+    @Inject
+    private SlingSettingsService settings;
 
     @PostConstruct
-    protected void initModel() {    	
-    	compHashCode = properties.get(Teaser.PN_TRACKING_OBJECT_ID,compHashCode);
-    	trackingEnabled = properties.get(Teaser.PN_TRACKING_ENABLED, trackingEnabled);
-    	super.initModel();
-      	id = String.valueOf(Math.abs(resource.getPath().hashCode()-1));   
-      	populateObjectId(); 
+    protected void initModel() {
+        compHashCode = properties.get(Teaser.PN_TRACKING_OBJECT_ID, compHashCode);
+        trackingEnabled = properties.get(Teaser.PN_TRACKING_ENABLED, trackingEnabled);
+        super.initModel();
+        hashCodeValue = String.valueOf(Math.abs(resource.getPath().hashCode() - 1));
+        runmode = settings.getRunModes();
+        if (runmode.contains(Externalizer.AUTHOR)) {
+            populateObjectId();
+        }
     }
-   
+
     @Override
-	protected void populateActions() {
+    protected void populateActions() {
         Resource actionsNode = resource.getChild(Teaser.NN_ACTIONS);
         if (actionsNode != null) {
-            for(Resource action : actionsNode.getChildren()) {
+            for (Resource action : actionsNode.getChildren()) {
                 actions.add(new ListItem() {
 
                     private ValueMap properties = action.getValueMap();
@@ -136,15 +145,18 @@ public class TeaserImpl extends com.adobe.cq.wcm.core.components.internal.models
                             return url;
                         }
                     }
+
                     @Nullable
                     @Override
-                    public String getAnalyticsData() {
-                    	analyticData = new StringBuilder();
-                    	counter++;
-                    	if(trackingEnabled && !compHashCode.isEmpty()){
-                    		analyticData = analyticData.append(OBJECT_ID).append(APOSTROPHE).append( compHashCode).append( SPACE).append(counter).append(APOSTROPHE).append(SEMICOLON);                	       
-                    	}
-                    	return analyticData.toString();
+                    public String getLinkTrackingCode() {
+                        linkTrackingCode = new StringBuilder();
+                        counter++;
+                        if (trackingEnabled && !compHashCode.isEmpty()) {
+                            linkTrackingCode = linkTrackingCode.append(OBJECT_ID).append(APOSTROPHE)
+                                    .append(compHashCode).append(StringUtils.SPACE).append(counter).append(APOSTROPHE)
+                                    .append(SEMICOLON);
+                        }
+                        return linkTrackingCode.toString();
                     }
                 });
             }
@@ -155,25 +167,27 @@ public class TeaserImpl extends com.adobe.cq.wcm.core.components.internal.models
     public List<ListItem> getActions() {
         return actions;
     }
-        
+
     @Override
-    public String getAnalyticsData() {
-    	analyticData = new StringBuilder() ;
-    	if(trackingEnabled && !compHashCode.isEmpty()){
-    		analyticData = analyticData.append(OBJECT_ID).append(APOSTROPHE).append(compHashCode).append(APOSTROPHE).append(SEMICOLON) ;	        
-    	}
-    	return analyticData.toString();
+    public String getLinkTrackingCode() {
+        linkTrackingCode = new StringBuilder();
+        if (trackingEnabled && !compHashCode.isEmpty()) {
+            linkTrackingCode = linkTrackingCode.append(OBJECT_ID).append(APOSTROPHE).append(compHashCode)
+                    .append(APOSTROPHE).append(SEMICOLON);
+        }
+        return linkTrackingCode.toString();
     }
-    
-    private void populateObjectId(){
-    	ModifiableValueMap map = resource.adaptTo(ModifiableValueMap.class);
-    	if (null != map && compHashCode.isEmpty() ) {
-    		map.put(Teaser.PN_TRACKING_OBJECT_ID, currentPage.getName()+ SPACE + component.getCellName()+ SPACE + id);
-    	}
-    	try {
-    		resourceResolver.commit();
-    	} catch (PersistenceException e) {
-    		LOGGER.error("Error occured while saving the objectId for {}", currentPage.getName(), e);
-    	}    	
+
+    private void populateObjectId() {
+        ModifiableValueMap map = resource.adaptTo(ModifiableValueMap.class);
+        if (null != map && compHashCode.isEmpty()) {
+            map.put(Teaser.PN_TRACKING_OBJECT_ID, currentPage.getName() + StringUtils.SPACE + component.getCellName()
+                    + StringUtils.SPACE + hashCodeValue);
+        }
+        try {
+            resourceResolver.commit();
+        } catch (PersistenceException e) {
+            LOGGER.error("Error occured while saving the objectId for {}", currentPage.getName(), e);
+        }
     }
 }
