+++ content/src/content/jcr_root/apps/core/wcm/components/image/v3/image/_cq_design_dialog/.content.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-  ~ Copyright 2017 Adobe Systems Incorporated
+  ~ Copyright 2019 Adobe Systems Incorporated
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -95,12 +95,13 @@
                                         checked="{Boolean}true"
                                         uncheckedValue="false"
                                         value="{Boolean}true"/>
-                                    <linktrackingEnabled
+                                    <linkTrackingEnabled
                                         jcr:primaryType="nt:unstructured"
                                         sling:resourceType="granite/ui/components/coral/foundation/form/checkbox"                                                
                                         fieldDescription="When clicked the link tracking option will be enabeld for user."
-                                        name="./linktrackingEnabled"
+                                        name="./linkTrackingEnabled"
                                         text="Enable Link Tracking"
+                                        checked="${not empty cqDesign.linkTrackingEnabled ? cqDesign.linkTrackingEnabled : false}"
                                         uncheckedValue="{Boolean}false"
                                         value="{Boolean}true"/>
                                     <widths
