+++ content/src/content/jcr_root/apps/core/wcm/components/image/v3/image/clientlibs/editor/js.txt
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright 2016 Adobe Systems Incorporated
+# Copyright 2019 Adobe Systems Incorporated
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
