+++ content/src/content/jcr_root/apps/core/wcm/components/teaser/v2/teaser/action.html
@@ -14,5 +14,5 @@
   ~ limitations under the License.
   ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~*/-->
 <sly data-sly-template.action="${@ action}">
-    <a class="cmp-teaser__action-link" href="${action.URL}" onClick = "${action.analyticsData @ context='html'}">${action.title}</a>
+    <a class="cmp-teaser__action-link" href="${action.URL}" onClick = "${action.linkTrackingCode @ context='html'}">${action.title}</a>
 </sly>
